feat(explore-search): Iterating on PR suggestions
