+++ static/app/views/explore/content.tsx
@@ -1,3 +1,4 @@
+import {useMemo} from 'react';
 import styled from '@emotion/styled';
 import type {Location} from 'history';
 
@@ -14,7 +15,10 @@ import {space} from 'sentry/styles/space';
 import useOrganization from 'sentry/utils/useOrganization';
 import usePageFilters from 'sentry/utils/usePageFilters';
 
+import {useGroupBys} from './hooks/useGroupBys';
+import {useResultMode} from './hooks/useResultsMode';
 import {useUserQuery} from './hooks/useUserQuery';
+import {useVisualizes} from './hooks/useVisualizes';
 import {ExploreCharts} from './charts';
 import {ExploreTables} from './tables';
 import {ExploreToolbar} from './toolbar';
@@ -26,6 +30,15 @@ interface ExploreContentProps {
 export function ExploreContent({}: ExploreContentProps) {
   const organization = useOrganization();
   const {selection} = usePageFilters();
+  const [groupBys] = useGroupBys();
+  const [visualizes] = useVisualizes();
+  const [resultsMode] = useResultMode();
+
+  const fields = useMemo(() => {
+    return [...groupBys, ...visualizes.flatMap(visualize => visualize.yAxes)].filter(
+      Boolean
+    );
+  }, [groupBys, visualizes]);
 
   const [userQuery, setUserQuery] = useUserQuery();
 
@@ -47,6 +60,8 @@ export function ExploreContent({}: ExploreContentProps) {
                   <DatePageFilter />
                 </PageFilterBar>
                 <SpanSearchQueryBuilder
+                  fields={fields}
+                  allowAggregateFunctions={resultsMode === 'aggregate'}
                   projects={selection.projects}
                   initialQuery={userQuery}
                   onSearch={setUserQuery}
